0f365d630b Localisation updates from https://translatewiki.net.
145d205b4c specials: Fix PHP Warning on Special:PasswordReset for crafted input
c60a5c4206 SECURITY: Escape newpage message in FeedUtils
db82379f65 installer: fix MySQL create user permissions check
20505ef627 INSTALL: Document requirement for bcmath/gmp on 32-bit systems
c2b5de68b3 Localisation updates from https://translatewiki.net.
f7a0d003c5 language: Avoid warning when 'namespaceGenderAliases' is null
723318ef91 http: Handle accept header with incomplete q
ab97f556b8 Update Pingback address
03d78bbd83 Localisation updates from https://translatewiki.net.
4ed3a3b051 filerepo: No exception on redirect without width in ThumbnailEntryPoint
5049b1c0d7 Localisation updates from https://translatewiki.net.
7f5acfeaca objectcache: Cast explicitly to integer
b7432b90c0 Localisation updates from https://translatewiki.net.
f418b5f3bf Localisation updates from https://translatewiki.net.
f8a4269751 tests: Use GLOB_BRACE in JsonSchemaAssertionTraitTest
3ba0c1b297 tests: Fix casing of MediaWiki in @covers
700a639140 FormatMetadata::formatFraction: Don't risk passing null to preg_match
cd0e100668 Localisation updates from https://translatewiki.net.
37efdff89b Link mw.Uri migration guide in docs and log warnings
7d44a728b6 title: Reset cached Title objects between tests
8e14ff1bc0 phpunit: Remove superfluous Title::clearCaches() calls
3ac4f81c18 Remove the hyphen as a legal search character for MySQL
9d3de690e0 Use anonymous user when creating named account from temp account
834c1a5407 Localisation updates from https://translatewiki.net.
a43f368b65 Update RfC links to bypass redirect
9ac90f7816 Improve BrokenRedirects display
e49931aca1 Sync up core repo with Parsoid
d46488212a Update wikimedia/parsoid to 0.20.3
994f95f757 Treat File::getShortDesc() as possibly unsafe HTML
d5e97d42ac Localisation updates from https://translatewiki.net.
20d15ad97f tests: Match deprecation message under php8.4 in DeprecationHelperTest
83d4898319 debug: Migrate E_USER_ERROR to throw Error in DeprecationHelper
8b7f9129c3 Setup: Switch vendor error from echo+E_USER_ERROR to echo+exit
78d252789c Setup: Update error message for composer dependencies check
f34ea0830a widget: Remove outdated try/catch wrapper from SpinnerWidget
50b0aadc6a phpunit: Remove unused trigger_error from TestLogger
a18e6ca5e6 ApiQueryRevisionsBase: Cast ctype_digit() param to string
5e27e3d4c9 Localisation updates from https://translatewiki.net.
bce2f8850d logger: Make log() methods return void
ac8a5b0b59 Drop PHP 7.4/8.0 support from master (forward-port from MW 1.42)
65c90be7ea Drop a few phan PhanImpossibleTypeComparison suppressions now we've dropped PHP 7.4
8a21d033f8 Clean up resource type and phan suppression in postgres code
82d3f6ac64 composer.json: Updated nikic/php-parser from ^4.10.2 to ^5.3.1
ed94cc0b64 structure tests: allow PHP 8.1 syntax and autoload enums
28f825fa8f composer.json: Updated nikic/php-parser from ^5.3.1 to ^5.5.0
38a0f5d5e9 Localisation updates from https://translatewiki.net.
81de7a814f RELEASE-NOTES-1.43: Update
4659cbcccc rdbms: fix table prefixing in "FOR UPDATE" clause generation in Postgres
7f1fa6f51f RefreshLinksJob: Check hastext before comparing HTML
db6013aa6c Api: Fix permission checks in action=compare
4fb27b070e htmlform: fix min/max validations on empty input in int/float fields
bbd7861a38 specials: SpecialTalkPage: Use config from request context
c3bdae9eca exception: Avoid service container init in exception handler
101639f81c exception: Skip use of HookRunner when not autoloaded
da8d30e0c6 [REST Sandbox] Remove SwaggerUI from MediaWiki Releases
4b59ce1f25 SECURITY: API: Escape i18n messages in action=feedcontributions
8a9ef65437 SECURITY: fix IP leak to unverified email
35edc6c2b1 SECURITY: apisandbox: Fix reflected XSS when invalid 'format' is provided
3340302f40 SECURITY: Do not treat autocreation as login for reauthentication
b1adf3c728 SECURITY: Allow extensions to supress the reauth flag on login
b2310f4736 SECURITY: Escape usernames in HTMLUserTextField validation errors
87ea0c739f SECURITY: Fix autoblocks visibility when bl_deleted=1
686589d207 SECURITY: BlockList: Hide rows containing suppressed users
71df9ed5b2 SECURITY: Fix leak of hidden usernames via autoblocks of those users
0af28b77ee Prep 1.43.2
923f7e6755 BlockListPagerTest: Fixup tests for pre REL1_44
c129ebd2d2 Localisation updates from https://translatewiki.net.
00290223cc composer.json: Replace away symfony php polyfills for PHP8/8.1
db289a0db6 Start 1.43.3

Depends-On: I5866b9b822b50904c2d8ceabcf49a8dfa97ab842

+++ includes/objectcache/SqlBagOStuff.php
@@ -1520,7 +1520,7 @@ class SqlBagOStuff extends MediumSpecificBagOStuff {
 					// `( 9 / 10 ) + ( 0.3 / 10 ) = 0.93`, or 93% done, overall.
 					$overallRatio = ( $progress['serversDone'] / $progress['serversTotal'] ) +
 						( $tablesDoneRatio / $progress['serversTotal'] );
-					( $progress['fn'] )( $overallRatio * 100 );
+					( $progress['fn'] )( (int)( $overallRatio * 100 ) );
 				}
 			} while ( $res->numRows() && $keysDeletedCount < $limit );
 		}
